Add test for parameters count metric
